[AMA-016] - Add sorting logic (#17)
* fix: create util dto

* update: add sortby and orderby param filters
